requirements.txt
pandas
numpy
openai
python-dotenv
pytest
openpyxl
xlrd
streamlit
matplotlib
seaborn
plotly
pandasql